% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CMRmeta.R
\name{CMRmeta}
\alias{CMRmeta}
\title{This function retrieves metadata from wml time series of gauged daily flow from the NRFA database.}
\usage{
CMRmeta(ID)
}
\arguments{
\item{ID}{station ID number(s), each number should be in the range [3002,236051].}
}
\value{
list composed of as many objects as in the list of station ID numbers. Each object can be accessed using their names or index (e.g. x[[1]], x[[2]], and so forth). Each object contains a data.frame.
}
\description{
Given the station ID number(s), this function retrieves data (time series in zoo format) and metadata.
}
\examples{
# One station
CMRmeta(3001)
# Multiple stations
# x <- CMRmeta(c(3001,3002,3003))
# View(x$ID3001)
# View(x[[1]])
}
\author{
Claudia Vitolo
}

