% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_flow.R
\name{convert_flow}
\alias{convert_flow}
\title{Convert flow from cumecs to mm/d}
\usage{
convert_flow(flowCumecs, catchmentArea)
}
\arguments{
\item{flowCumecs}{This is the flow time series in cumecs (m3/s)}

\item{catchmentArea}{This is the catchment are in Km2.}
}
\value{
Flow time series in mm/d
}
\description{
This function converts flow time series from cumecs (m3/s) to mm/d by dividing the flow by the catchment area and converting it to mm/day.
}
\examples{
\dontrun{
  convert_flow(30, 2)
}

}
