% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GenerateMap.R
\name{GenerateMap}
\alias{GenerateMap}
\title{Generate map of gauging stations.}
\usage{
GenerateMap(selectedStationSummary)
}
\arguments{
\item{selectedStationSummary}{This is the data.frame containing at least a column called "gridReference" (in which OS Grid references are stored) with 1 row for each station. Alternatively the table could contain 2 columns called "Latitude" and "Longitude" containing the coordinates of stations in WGS84, with 1 row for each station.}
}
\value{
displays a map
}
\description{
This function takes as input a table of UK NRFA stations (output of getStationSummary() function) and generates a map using leaflet javascript library.
}
\examples{
# bbox <- list(lonMin=-3.82, lonMax=-3.63, latMin=52.43, latMax=52.52)
# myStations <- NRFA_Catalogue(bbox)
# GenerateMap(myStations)
}
\author{
Claudia Vitolo
}

