% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getStationSummary}
\alias{getStationSummary}
\title{List of stations from UK NRFA}
\usage{
getStationSummary(LonMin = -10, LonMax = 10, LatMin = 48, LatMax = 62,
  mdCol = NULL, mdVal = NULL, minRec = NULL)
}
\arguments{
\item{LonMin}{Minimum longitude of bounding box}

\item{LonMax}{Maximum longitude of bounding box}

\item{LatMin}{Minimum latitude of bounding box}

\item{LatMax}{Maximum latitude of bounding box}

\item{mdCol}{name of column to filter}

\item{mdVal}{string to search in mdCol}

\item{minRec}{minimum number of recording years}
}
\value{
data.frame with list of stations and related metadata
}
\description{
This function pulls the list of stations (and related metadata), falling within a given bounding box, from the CEH National River Flow Archive website.
}
\details{
coordinates of bounding box are required in WGS84 (EPSG: 4326). If BB coordinates are missing, the function returns the list corresponding to the maximum extent of the network.
}
\examples{
x <- getStationSummary() # this returns all the stations in the network
x <- getStationSummary(LonMin=-1, LonMax=1, LatMin=49, LatMax=51) # this returns 31 stations
}
\author{
Claudia Vitolo
}

