% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NRFA_TS.R
\name{NRFA_TS}
\alias{NRFA_TS}
\title{This function retrieves time series of gauged daily flow from the NRFA database.}
\usage{
NRFA_TS(ID)
}
\arguments{
\item{ID}{station ID number(s), each number should be in the range [3002,236051].}
}
\value{
list composed of as many objects as in the list of station ID numbers. Each object can be accessed as x[[1]], x[[2]], and so forth. Each object contains a list of other two objects: data and metadata.
}
\description{
Given the station ID number(s), this function retrieves data (time series in zoo format) and metadata.
}
\examples{
# One station
NRFA_TS(3001)
# Multiple stations
# NRFA_TS(c(3001,3002,3003)); plot(x[[1]]$data)
}
\author{
Claudia Vitolo
}

