% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRFA_TSdata.R
\name{NRFA_TSdata}
\alias{NRFA_TSdata}
\title{This function retrieves data from wml time series of gauged daily flow from the NRFA database.}
\usage{
NRFA_TSdata(ID)
}
\arguments{
\item{ID}{station ID number(s), each number should be in the range [3002,236051].}
}
\value{
list composed of as many objects as in the list of station ID numbers. Each object can be accessed using their names or index (e.g. x[[1]], x[[2]], and so forth). Each object contains a zoo time series.
}
\description{
Given the station ID number(s), this function retrieves data (time series in zoo format) and metadata.
}
\examples{
# One station
NRFA_TSdata(3001)
# Multiple stations
# x <- NRFA_TSdata(c(3001,3002,3003))
# plot(x$ID3001)
# plot(x[[1]])

}
\author{
Claudia Vitolo
}

