% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRFA_TSmetadata.R
\name{NRFA_TSmetadata}
\alias{NRFA_TSmetadata}
\title{This function retrieves metadata from wml time series of gauged daily flow from the NRFA database.}
\usage{
NRFA_TSmetadata(ID)
}
\arguments{
\item{ID}{station ID number(s), each number should be in the range [3002,236051].}
}
\value{
list composed of as many objects as in the list of station ID numbers. Each object can be accessed using their names or index (e.g. x[[1]], x[[2]], and so forth). Each object contains a data.frame.
}
\description{
Given the station ID number(s), this function retrieves data (time series in zoo format) and metadata.
}
\examples{
# One station
NRFA_TSmetadata(3001)
# Multiple stations
# x <- NRFA_TSmetadata(c(3001,3002,3003))
# View(x$ID3001)
# View(x[[1]])

}
\author{
Claudia Vitolo
}

