% Generated by roxygen2 (4.0.1): do not edit by hand
\name{osg2latlon1}
\alias{osg2latlon1}
\title{Converts easting/northing to latitude/longitude.}
\usage{
osg2latlon1(en)
}
\arguments{
\item{en}{this is a vector containing the easting and northing coordinates.}
}
\value{
vector made of two elements: the latitude and longitude.
}
\description{
This function converts easting and northing coordinates (UK National Grid, epsg:27700) and converts them to longitute and latitude (WGS84, epsg:4326). It uses rgadal and sp functionalities for coordinates transformations.
}
\examples{
# single entry
osg2latlon(c(572200,121300))

# use result from osgparse() function
osg2latlon(osgparse("TQ722213"))

# multiple entries & use result from osgparse() function
osg2latlon(osgparse(c("SN831869","SN829838","SN824853","SN824842","SN826854")))
}
\author{
Claudia Vitolo
}
\references{
This function is based on the following forum post: https://stat.ethz.ch/pipermail/r-sig-geo/2010-November/010141.htm
}

