% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend.R
\name{plot_trend}
\alias{plot_trend}
\title{Plot trend}
\usage{
plot_trend(df, column_name)
}
\arguments{
\item{df}{Data frame containing at least 4 column:
lat (latitude), lon (longitude), slope and an additional user-defined column
\code{column_name}.}

\item{column_name}{name of the column to use for grouping the results.}
}
\value{
Two plots, side-by-side, the first showing the distribution of the
trend over a map, based on the slope of the linear model that describes the
trend. The second plot shows a boxplot of the slope grouped based on the
column Region. Region and slope can be user-defined.
}
\description{
This function plots a previously calculated trend.
}
\examples{
\dontrun{
  plot_trend(df, Region)
}
}
