% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{print.nrfa_api}
\alias{print.nrfa_api}
\title{print.nrfa_api}
\usage{
\method{print}{nrfa_api}(x, ...)
}
\arguments{
\item{x}{an nrfa call}

\item{...}{additional items, not really used}
}
\description{
Not to be used by user. Internal function used to print error in nrfaapi.
}
