% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnrfa-package.R
\docType{package}
\name{rnrfa-package}
\alias{rnrfa-package}
\title{UK National River Flow Archive data from R}
\description{
rnrfa: UK National River Flow Archive Data from R.
}
\details{
Utility functions to retrieve data from the UK National River Flow Archive
(http://nrfa.ceh.ac.uk/). The package contains R wrappers to the UK NRFA data
temporary-API. There are functions to retrieve stations falling in a bounding
box, to generate a map and extracting time series and general information.
}
