% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_ids.R
\name{station_ids}
\alias{station_ids}
\title{List of stations identification numbers from UK NRFA}
\usage{
station_ids()
}
\value{
vector integer identification numbers (one for each station)
}
\description{
This function pulls the list of station identification numbers.
}
\examples{
\dontrun{
  # Retrieve all the stations ids
  x <- station_ids()
}

}
\author{
Claudia Vitolo
}
