% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ToJSONarray.R
\name{ToJSONarray}
\alias{ToJSONarray}
\title{Transforms a data.frame into a JSON array in a format compatible with D3.js and Protovis
(source by: http://theweiluo.wordpress.com/2011/09/30/r-to-json-for-d3-js-and-protovis/)}
\usage{
ToJSONarray(dtf)
}
\arguments{
\item{dtf}{this is a data.frame}
}
\value{
JSON array in a format compatible with D3.js and Protovis
}
\description{
This function transforms a R data frame into a JSON array such that each row in R becomes an JSON object in the array.
}
\details{
The rjson/RJSONIO packages convert a data frame into a JSON hashmap, in which each column of the data frame becomes a named array.
}
\examples{
# ToJSONarray(dtf)
}
\author{
Claudia Vitolo
}

