% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findInfo}
\alias{findInfo}
\title{Extract info from the waterml2.}
\usage{
findInfo(myList)
}
\arguments{
\item{myList}{this is a nested list which comes from the conversion of waterml2 data}
}
\value{
named vector containing the following information: stationName, Latitude, Longitude, typeOfMeasurement
}
\description{
This function retrieves the most important info from the waterml2 file given as a list.
}
\author{
Claudia Vitolo
}

