% Generated by roxygen2 (4.0.2): do not edit by hand
\name{FindTS}
\alias{FindTS}
\title{This function retrieves time series of gauged daily flow from the NRFA database.}
\usage{
FindTS(myList)
}
\arguments{
\item{myList}{this is a nested list which comes from the conversion of waterml2 data.}
}
\value{
time series of class zoo
}
\description{
This function retrieves time series of gauged daily flow from the NRFA database and stores it in zoo format.
}
\author{
Claudia Vitolo
}

