% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generateMap}
\alias{generateMap}
\title{Generate map of UK NRFA stations.}
\usage{
generateMap(StationSummary)
}
\arguments{
\item{StationSummary}{This is the data.frame containing at least a column called "gridReference" (in which OS Grid references are stored) with 1 row for each station. Alternatively the table could contain 2 columns called "Latitude" and "Longitude" containing the coordinates of stations in WGS84, with 1 row for each station.}
}
\value{
displays a map
}
\description{
This function takes as input a table of UK NRFA stations (output of getStationSummary() function) and generates a map using leaflet javascript library.
}
\examples{
# myMap <- getStationSummary(LonMin=-1, LonMax=1, LatMin=51, LatMax=49)
# generateMap(myMap)
}
\author{
Claudia Vitolo
}

