% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GenerateMap}
\alias{GenerateMap}
\title{Generate map of UK NRFA stations.}
\usage{
GenerateMap(selectedStationSummary)
}
\arguments{
\item{selectedStationSummary}{This is the data.frame containing at least a column called "gridReference" (in which OS Grid references are stored) with 1 row for each station. Alternatively the table could contain 2 columns called "Latitude" and "Longitude" containing the coordinates of stations in WGS84, with 1 row for each station.}
}
\value{
displays a map
}
\description{
This function takes as input a table of UK NRFA stations (output of getStationSummary() function) and generates a map using leaflet javascript library.
}
\examples{
# myStations <- GetStationSummary(lonMin=-1, lonMax=1, latMin=51, latMax=49)
# GenerateMap(myStations)
}
\author{
Claudia Vitolo
}

