% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SearchNRFA}
\alias{SearchNRFA}
\title{This function retrieves time series of gauged daily flow from the NRFA database.}
\usage{
SearchNRFA(stationID)
}
\arguments{
\item{stationID}{Station ID number, it should be in the range [3002,236051].}
}
\value{
time series of class zoo
}
\description{
Given the station ID number, this function retrieves the time series in zoo format.
}
\examples{
SearchNRFA(3002)
}
\author{
Claudia Vitolo
}

